Split out `assertions` module
